pyproject.toml
[project]
name = "tach"
version = "0.6.3"
authors = [
  { name="Caelean Barnes", email="caeleanb@gmail.com" },
  { name="Evan Doyle", email="evanmdoyle@gmail.com" },
]
description = "A Python tool to maintain a modular package architecture."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml~=6.0",
    "pydantic~=2.0",
    "rich~=13.0",
    "prompt-toolkit~=3.0",
    "stdlib-list>=0.10.0; python_version < '3.10'",
    "eval-type-backport>=0.2.0; python_version < '3.10'"
]
keywords = ['python', 'module', 'package', 'guard', 'enforcement', 'boundary', 'enforcer', 'domain', 'architecture']


[project.urls]
Homepage = "https://github.com/gauge-sh/tach"
Issues = "https://github.com/gauge-sh/tach/issues"

[tool.ruff]
src = ["python"]
exclude = ["python/tests/example"]
target-version = "py38"
lint.extend-select = ["I", "TCH", "UP"]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = [
    "pydantic.BaseModel",
    "pydantic.RootModel",
]
runtime-evaluated-decorators = [
    "pydantic.dataclasses.dataclass",
    "dataclasses.dataclass",
]
exempt-modules = ["typing", "typing_extensions"]

[tool.pyright]
include = ["python/tach"]
exclude = ["**/__pycache__", ".venv"]
strict = ["python/tach"]
executionEnvironments = [
  { root = "python" }
]
pythonVersion = "3.8"

[build-system]
requires = ["maturin>=1.5,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"
module-name = "tach.extension"
features = ["pyo3/extension-module"]

[project.scripts]
tach = "tach.start:start"
